# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Reisetagebuch API Draft",
        "description": "This is a draft version of the Reisetagebuch API documentation. The API is still under development and may change in future releases. Use at your own risk!",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "0.1.0-draft",
        "x-logo": {
            "url": "https://raw.githubusercontent.com/HerrLevin/reisetagebuch/refs/heads/main/public/assets/logo.webp"
        }
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/location/prefetch": {
            "get": {
                "tags": [
                    "Location"
                ],
                "summary": "Prefetch location",
                "description": "Prefetch location data and optionally store user history",
                "operationId": "prefetchLocation",
                "parameters": [
                    {
                        "name": "latitude",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "longitude",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/location/request-location": {
            "get": {
                "tags": [
                    "Location"
                ],
                "summary": "Get recent request location",
                "description": "Return a recent location matching the requested coordinates",
                "operationId": "getRecentRequestLocation",
                "parameters": [
                    {
                        "name": "latitude",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "longitude",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "App\\Dto\\RequestLocationDto"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Location not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "App\\Dto\\ErrorDto"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/locations/nearby": {
            "get": {
                "tags": [
                    "Location"
                ],
                "summary": "Search locations",
                "description": "Search for locations near a point or by query",
                "operationId": "searchLocations",
                "parameters": [
                    {
                        "name": "latitude",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "longitude",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "query",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/LocationDto"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/geocode": {
            "get": {
                "tags": [
                    "Location"
                ],
                "summary": "Geocode",
                "description": "Geocode a query using configured providers",
                "operationId": "geocode",
                "parameters": [
                    {
                        "name": "query",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "provider",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "latitude",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "longitude",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/MotisGeocodeResponseEntry"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/locations/history": {
            "get": {
                "tags": [
                    "Location"
                ],
                "summary": "Location history",
                "description": "Return location history for the authenticated user for a given day",
                "operationId": "locationHistory",
                "parameters": [
                    {
                        "name": "when",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LocationHistoryDto"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/locations/departures": {
            "get": {
                "tags": [
                    "Location"
                ],
                "summary": "Departures",
                "description": "Return departures for a given point or station identifier",
                "operationId": "departures",
                "parameters": [
                    {
                        "name": "latitude",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "longitude",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "identifier",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "when",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    },
                    {
                        "name": "modes",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/components/schemas/TransportMode"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DeparturesResponseDto"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/locations/stopovers": {
            "get": {
                "tags": [
                    "Location"
                ],
                "summary": "Stopovers",
                "description": "Return stopovers for a given trip start",
                "operationId": "stopovers",
                "parameters": [
                    {
                        "name": "tripId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "startId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "startTime",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StopoversResponseDto"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DeparturesResponseDto": {
                "description": "Data Transfer Object for Departures Response",
                "required": [
                    "departures",
                    "modes",
                    "requestTime",
                    "requestIdentifier",
                    "requestLatitude",
                    "requestLongitude"
                ],
                "properties": {
                    "departures": {
                        "$ref": "#/components/schemas/DeparturesDto"
                    },
                    "modes": {
                        "description": "Filter modes applied to the request",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TransportMode"
                        }
                    },
                    "requestTime": {
                        "description": "Time when the request was made",
                        "type": "string",
                        "format": "date-time"
                    },
                    "requestIdentifier": {
                        "description": "Unique identifier for the request",
                        "type": "string",
                        "nullable": true
                    },
                    "requestLatitude": {
                        "description": "Latitude of the request location",
                        "type": "number",
                        "format": "float"
                    },
                    "requestLongitude": {
                        "description": "Longitude of the request location",
                        "type": "number",
                        "format": "float"
                    }
                },
                "type": "object"
            },
...
            "LocationHistoryDto": {
                "description": "Data Transfer Object for Location History",
                "required": [
                    "locations",
                    "trips"
                ],
                "properties": {
                    "locations": {
                        "description": "Collection of location history entries",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/LocationHistoryEntryDto"
                        }
                    },
                    "trips": {
                        "description": "Collection of trip history entries",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TripHistoryEntryDto"
                        }
                    }
                },
                "type": "object"
            },
            "MotisGeocodeResponseEntry": {
                "required": [
                    "type",
                    "tokens",
                    "name",
                    "identifier",
                    "lat",
                    "lon",
                    "score",
                    "id",
                    "level",
                    "street",
                    "houseNumber",
                    "zip",
                    "areas",
                    "identifiers"
                ],
                "properties": {
                    "type": {
                        "$ref": "#/components/schemas/MotisLocationType"
                    },
                    "tokens": {
                        "description": "An array of tokens representing parts of the location name or identifier",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "name": {
                        "description": "The name of the location",
                        "type": "string"
                    },
                    "identifier": {
                        "description": "A unique identifier for the location",
                        "type": "string"
                    },
                    "id": {
                        "description": "The ID of the location",
                        "type": "string",
                        "nullable": true
                    },
                    "lat": {
                        "description": "The latitude of the location",
                        "type": "number",
                        "format": "float"
                    },
                    "lon": {
                        "description": "The longitude of the location",
                        "type": "number",
                        "format": "float"
                    },
                    "level": {
                        "description": "The level or floor of the location, if applicable",
                        "type": "string",
                        "nullable": true
                    },
                    "street": {
                        "description": "The street name of the location, if applicable",
                        "type": "string",
                        "nullable": true
                    },
                    "houseNumber": {
                        "description": "The house number of the location, if applicable",
                        "type": "string",
                        "nullable": true
                    },
                    "zip": {
                        "description": "The ZIP code of the location, if applicable",
                        "type": "string",
                        "nullable": true
                    },
                    "areas": {
                        "description": "An array of area names or identifiers associated with the location",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "score": {
                        "description": "The confidence score of the geocode result",
                        "type": "number",
                        "format": "float"
                    },
                    "identifiers": {
                        "description": "An array of location identifiers",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/LocationIdentifierDto"
                        }
                    }
                },
                "type": "object"
            },
...
            "TransportMode": {
                "description": "Enumeration of transport modes",
                "type": "string",
                "enum": [
                    "WALK",
                    "BIKE",
                    "RENTAL",
                    "CAR",
                    "CAR_PARKING",
                    "CAR_DROPOFF",
                    "ODM",
                    "FLEX",
                    "TRANSIT",
                    "TRAM",
                    "SUBWAY",
                    "FERRY",
                    "AIRPLANE",
                    "SUBURBAN",
                    "BUS",
                    "COACH",
                    "RAIL",
                    "HIGHSPEED_RAIL",
                    "LONG_DISTANCE",
                    "NIGHT_RAIL",
                    "REGIONAL_FAST_RAIL",
                    "REGIONAL_RAIL",
                    "CABLE_CAR",
                    "FUNICULAR",
                    "AERIAL_LIFT",
                    "OTHER",
                    "AREAL_LIFT",
                    "METRO"
                ]
            },
...
            "LocationDto": {
                "description": "Location Data Object",
                "required": [
                    "id",
                    "name",
                    "latitude",
                    "longitude",
                    "distance",
                    "tags",
                    "identifiers"
                ],
                "properties": {
                    "id": {
                        "description": "Location ID",
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "description": "Name of the location",
                        "type": "string"
                    },
                    "latitude": {
                        "description": "Latitude of the location",
                        "type": "number",
                        "format": "float"
                    },
                    "longitude": {
                        "description": "Longitude of the location",
                        "type": "number",
                        "format": "float"
                    },
                    "distance": {
                        "description": "Distance to the location in meters",
                        "type": "integer",
                        "nullable": true
                    },
                    "identifiers": {
                        "description": "List of location identifiers",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/LocationIdentifierDto"
                        }
                    },
                    "tags": {
                        "description": "List of location tags",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/LocationTagDto"
                        }
                    }
                },
                "type": "object"
            },
...
            "LocationIdentifierDto": {
                "description": "Location Identifier Data Object",
                "required": [
                    "type",
                    "origin",
                    "identifier"
                ],
                "properties": {
                    "type": {
                        "description": "Type of the location identifier",
                        "type": "string"
                    },
                    "origin": {
                        "description": "Origin of the location identifier",
                        "type": "string"
                    },
                    "identifier": {
                        "description": "The location identifier value",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    },
...
    "security": [
        {
            "passport": []
        }
    ]
}